test(components):
Jira issue: VEGA-523
