+++ src/components/Table2/TableCell2/validate-value.test.ts
@@ -1,0 +1,12 @@
+import { validateValue } from './validateValue';
+
+describe('validateValue tableCell2', () => {
+  test('input in tableCell2', () => {
+    const input = '111,,.3.,44.';
+    const input2 = '111,,.3.,44.77';
+    const output = '111344.00';
+    const output2 = '111344.77';
+    expect(validateValue(input)).toEqual(output);
+    expect(validateValue(input2)).toEqual(output2);
+  });
+});
